Rekonfigurer prometheus
