+++ src/main/kotlin/no/nav/syfo/metrics/AppMetrics.kt
@@ -10,40 +10,45 @@ import io.micrometer.prometheus.PrometheusMeterRegistry
 import io.prometheus.client.CollectorRegistry
 import io.prometheus.client.hotspot.DefaultExports
 
-const val APP_METRICS_NS = "esyfovarsel"
+const val METRICS_NS = "esyfovarsel"
 
-const val ERROR_IN_PLANNER = "${APP_METRICS_NS}_error_in_planner"
-const val ERROR_IN_PARSING = "${APP_METRICS_NS}_error_in_parser"
-const val ERROR_IN_PROCESSING = "${APP_METRICS_NS}_error_in_processing"
-const val MER_VEILEDNING_PLANNED = "${APP_METRICS_NS}_mer_veiledning_planned"
-const val AKTIVITETSKRAV_PLANNED = "${APP_METRICS_NS}_aktivitetskrav_planned"
+const val ERROR_IN_PLANNER = "${METRICS_NS}_error_in_planner"
+const val ERROR_IN_PARSING = "${METRICS_NS}_error_in_parser"
+const val ERROR_IN_PROCESSING = "${METRICS_NS}_error_in_processing"
+const val MER_VEILEDNING_PLANNED = "${METRICS_NS}_mer_veiledning_planned"
+const val AKTIVITETSKRAV_PLANNED = "${METRICS_NS}_aktivitetskrav_planned"
 
-val APP_METRICS_REGISTRY = PrometheusMeterRegistry(PrometheusConfig.DEFAULT, CollectorRegistry.defaultRegistry, Clock.SYSTEM)
+const val MER_VEILEDNING_NOTICE_SENT = "${METRICS_NS}_mer_veiledning_notice_sent"
+const val AKTIVITETSKRAV_NOTICE_SENT = "${METRICS_NS}_aktivitetskrav_notice_sent"
+const val NOTICE_SENT = "${METRICS_NS}_notice_sent"
+
+
+val METRICS_REGISTRY = PrometheusMeterRegistry(PrometheusConfig.DEFAULT, CollectorRegistry.defaultRegistry, Clock.SYSTEM)
 
 val COUNT_ERROR_IN_PLANNER: Counter = Counter
     .builder(ERROR_IN_PLANNER)
     .description("Counts the number of all errors in planner")
-    .register(APP_METRICS_REGISTRY)
+    .register(METRICS_REGISTRY)
 
 val COUNT_ERROR_IN_PARSING: Counter = Counter
     .builder(ERROR_IN_PARSING)
     .description("Counts the number of all errors in parsing")
-    .register(APP_METRICS_REGISTRY)
+    .register(METRICS_REGISTRY)
 
 val COUNT_ERROR_IN_PROCESSING: Counter = Counter
     .builder(ERROR_IN_PROCESSING)
     .description("Counts the number of all errors in processing")
-    .register(APP_METRICS_REGISTRY)
+    .register(METRICS_REGISTRY)
 
 val COUNT_MER_VEILEDNING_PLANNED: Counter = Counter
     .builder(MER_VEILEDNING_PLANNED)
     .description("Counts the number of planned notice of type Mer veiledning")
-    .register(APP_METRICS_REGISTRY)
+    .register(METRICS_REGISTRY)
 
 val COUNT_AKTIVITETSKRAV_PLANNED: Counter = Counter
     .builder(AKTIVITETSKRAV_PLANNED)
     .description("Counts the number of planned notice of type Aktivitetskrav")
-    .register(APP_METRICS_REGISTRY)
+    .register(METRICS_REGISTRY)
 
 
 fun tellFeilIPlanner() {
@@ -58,6 +63,31 @@ fun tellFeilIProsessering() {
     COUNT_ERROR_IN_PROCESSING.increment()
 }
 
+val COUNT_MER_VEILEDNING_NOTICE_SENT: Counter = Counter
+    .builder(MER_VEILEDNING_NOTICE_SENT)
+    .description("Counts the number of Mer veiledning notice sent")
+    .register(METRICS_REGISTRY)
+
+val COUNT_AKTIVITETSKRAV_NOTICE_SENT: Counter = Counter
+    .builder(AKTIVITETSKRAV_NOTICE_SENT)
+    .description("Counts the number of Aktivitetskrav notice sent")
+    .register(METRICS_REGISTRY)
+
+val COUNT_ALL_NOTICE_SENT: Counter = Counter
+    .builder(NOTICE_SENT)
+    .description("Counts the number of all types of notice sent")
+    .register(METRICS_REGISTRY)
+
+fun tellMerVeiledningVarselSendt(varslerSendt: Int) {
+    COUNT_ALL_NOTICE_SENT.increment(varslerSendt.toDouble())
+    COUNT_MER_VEILEDNING_NOTICE_SENT.increment(varslerSendt.toDouble())
+}
+
+fun tellAktivitetskravVarselSendt(varslerSendt: Int) {
+    COUNT_ALL_NOTICE_SENT.increment(varslerSendt.toDouble())
+    COUNT_AKTIVITETSKRAV_NOTICE_SENT.increment(varslerSendt.toDouble())
+}
+
 fun tellMerVeiledningPlanlagt() {
     COUNT_MER_VEILEDNING_PLANNED.increment()
 }
@@ -66,12 +96,10 @@ fun tellAktivitetskravPlanlagt() {
     COUNT_AKTIVITETSKRAV_PLANNED.increment()
 }
 
-
 fun Routing.registerPrometheusApi() {
     DefaultExports.initialize()
 
     get("/prometheus") {
-        call.respondText(APP_METRICS_REGISTRY.scrape())
+        call.respondText(METRICS_REGISTRY.scrape())
     }
 }
-
